Add quiz manager script that hides and shows a question in the VR portion when the question is due

+++ UnityProject/Assets/Scripts/Question.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Question
 {
     // Time in seconds
-    public int time;
+    public float time;
 
     // Title of the question / test of the question
     public string title;
